pyproject.toml
[project]
name = "nonrigid-icp-ed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "loguru>=0.7.3",
    "numpy>=2.2.6",
    "omegaconf>=2.3.0",
    "open3d>=0.19.0",
    "torch==2.9.0",
    "torchvision==0.24.0",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = " https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
